Move `PartyClownCar` to @scene/token-document
Also clarify some type-only imports
