+++ src/module/documents.ts
@@ -1,7 +1,12 @@
-export { AmbientLightDocumentPF2e, MeasuredTemplateDocumentPF2e, ScenePF2e, TokenDocumentPF2e } from "@scene";
-export { ActiveEffectPF2e } from "./active-effect.ts";
+// Client documents
 export { ActorPF2e } from "./actor/base.ts";
 export { ChatMessagePF2e } from "./chat-message/document.ts";
 export { ItemPF2e } from "./item/base.ts";
 export { MacroPF2e } from "./macro.ts";
 export { UserPF2e } from "./user/document.ts";
+
+// Embedded documents
+export { ActiveEffectPF2e } from "./active-effect.ts";
+
+// Canvas documents
+export { AmbientLightDocumentPF2e, MeasuredTemplateDocumentPF2e, ScenePF2e, TokenDocumentPF2e } from "@scene";
