+++ src/module/scene/index.ts
@@ -1,5 +1,7 @@
 export { ScenePF2e } from "./document.ts";
+
+// Embedded Documents
 export { AmbientLightDocumentPF2e } from "./ambient-light-document.ts";
-export { TileDocumentPF2e } from "./tile-document.ts";
-export { TokenConfigPF2e, TokenDocumentPF2e } from "./token-document/index.ts";
 export { MeasuredTemplateDocumentPF2e } from "./measured-template-document.ts";
+export { TileDocumentPF2e } from "./tile-document.ts";
+export { TokenDocumentPF2e } from "./token-document/index.ts";
