+++ src/scripts/register-sheets.ts
@@ -3,14 +3,15 @@ import { FamiliarSheetPF2e } from "@actor/familiar/sheet.ts";
 import { HazardSheetPF2e } from "@actor/hazard/sheet.ts";
 import { LootSheetPF2e } from "@actor/loot/sheet.ts";
 import { NPCSheetPF2e, SimpleNPCSheet } from "@actor/npc/sheet.ts";
+import { PartySheetPF2e } from "@actor/party/sheet.ts";
 import { VehicleSheetPF2e } from "@actor/vehicle/sheet.ts";
-import { ItemSheetPF2e } from "@item/sheet/base.ts";
 import { ActionSheetPF2e } from "@item/ability/sheet.ts";
 import { AfflictionSheetPF2e } from "@item/affliction/sheet.ts";
 import { AncestrySheetPF2e } from "@item/ancestry/sheet.ts";
 import { ArmorSheetPF2e } from "@item/armor/sheet.ts";
 import { BackgroundSheetPF2e } from "@item/background/sheet.ts";
 import { BookSheetPF2e } from "@item/book/sheet.ts";
+import { CampaignFeatureSheetPF2e } from "@item/campaign-feature/sheet.ts";
 import { ClassSheetPF2e } from "@item/class/sheet.ts";
 import { ConsumableSheetPF2e } from "@item/consumable/sheet.ts";
 import { ContainerSheetPF2e } from "@item/container/sheet.ts";
@@ -23,15 +24,15 @@ import { KitSheetPF2e } from "@item/kit/sheet.ts";
 import { MeleeSheetPF2e } from "@item/melee/sheet.ts";
 import { PhysicalItemSheetPF2e } from "@item/physical/sheet.ts";
 import { PHYSICAL_ITEM_TYPES } from "@item/physical/values.ts";
+import { ItemSheetPF2e } from "@item/sheet/base.ts";
 import { SpellSheetPF2e } from "@item/spell/index.ts";
 import { TreasureSheetPF2e } from "@item/treasure/sheet.ts";
 import { WeaponSheetPF2e } from "@item/weapon/sheet.ts";
 import { JournalSheetPF2e, JournalTextTinyMCESheetPF2e } from "@module/journal-entry/sheet.ts";
 import { UserConfigPF2e } from "@module/user/sheet.ts";
-import { TokenConfigPF2e, TokenDocumentPF2e } from "@scene/index.ts";
 import { SceneConfigPF2e } from "@scene/sheet.ts";
-import { PartySheetPF2e } from "@actor/party/sheet.ts";
-import { CampaignFeatureSheetPF2e } from "@item/campaign-feature/sheet.ts";
+import { TokenDocumentPF2e } from "@scene/token-document/document.ts";
+import { TokenConfigPF2e } from "@scene/token-document/sheet.ts";
 
 export function registerSheets(): void {
     const sheetLabel = game.i18n.localize("PF2E.SheetLabel");
